+++ src/Local_Development/Extras.php
@@ -85,26 +85,53 @@ class Extras extends Settings {
 			'local_dev_extras',
 			'local_dev_extras',
 			[
-				'id'   => 'local_servers',
-				'type' => 'extras',
-				'name' => esc_html( 'Enable Local Git Servers (192.168.x.x)', 'local-development' ),
+				'id'    => 'local_servers',
+				'type'  => 'extras',
+				'name'  => esc_html__( 'Enable Local Git Servers (192.168.x.x)', 'local-development' ),
+				'class' => $this->is_localhost() ? '' : 'hidden',
 			]
 		);
 
-		if ( version_compare( get_bloginfo( 'version' ), '5.2', '>=' ) ) {
-			add_settings_field(
-				'fatal_error_handler',
-				null,
-				[ $this, 'token_callback_checkbox' ],
-				'local_dev_extras',
-				'local_dev_extras',
-				[
-					'id'   => 'bypass_fatal_error_handler',
-					'type' => 'extras',
-					'name' => esc_html__( 'Bypass WordPress 5.1 WSOD protection.', 'local-development' ),
-				]
-			);
-		}
+		add_settings_field(
+			'fatal_error_handler',
+			null,
+			[ $this, 'token_callback_checkbox' ],
+			'local_dev_extras',
+			'local_dev_extras',
+			[
+				'id'    => 'bypass_fatal_error_handler',
+				'type'  => 'extras',
+				'name'  => esc_html__( 'Bypass WordPress 5.1 WSOD protection.', 'local-development' ),
+				'class' => version_compare( get_bloginfo( 'version' ), '5.2', '>=' ) ? '' : 'hidden',
+			]
+		);
+
+		add_settings_field(
+			'git_host_icons',
+			null,
+			[ $this, 'token_callback_checkbox' ],
+			'local_dev_extras',
+			'local_dev_extras',
+			[
+				'id'   => 'disable_git_icons',
+				'type' => 'extras',
+				'name' => esc_html__( 'Disable Git Host icons.', 'local-development' ),
+			]
+		);
+
+		add_settings_field(
+			'adminbar_visual_feedback',
+			null,
+			[ $this, 'token_callback_checkbox' ],
+			'local_dev_extras',
+			'local_dev_extras',
+			[
+				'id'    => 'disable_admin_bar_visual_feedback',
+				'type'  => 'extras',
+				'name'  => esc_html__( 'Disable custom Admin Bar styles for localhost.', 'local-development' ),
+				'class' => $this->is_localhost() ? '' : 'hidden',
+			]
+		);
 	}
 
 	/**
@@ -145,6 +172,10 @@ class Extras extends Settings {
 		if ( isset( self::$options['extras']['bypass_fatal_error_handler'] ) ) {
 			add_filter( 'wp_fatal_error_handler_enabled', '__return_false' );
 		}
+		if ( $this->is_localhost() && ! isset( static::$options['extras']['disable_admin_bar_visual_feedback'] ) ) {
+			add_action( 'admin_head', [ $this, 'custom_local_admin_bar_css' ] ); // on backend area.
+			add_action( 'wp_head', [ $this, 'custom_local_admin_bar_css' ] ); // on frontend area.
+		}
 	}
 
 	/**
@@ -179,5 +210,63 @@ class Extras extends Settings {
 			10,
 			2
 		);
+	}
+
+	/**
+	 * Check if this a local WordPress instance.
+	 */
+	private function is_localhost() {
+		// phpcs:ignore WordPress.Security.ValidatedSanitizedInput
+		$server_addr = isset( $_SERVER['REMOTE_ADDR'] ) ? $_SERVER['REMOTE_ADDR'] : null;
+		return in_array( $server_addr, [ '127.0.0.1', '::1' ], true );
+	}
+
+	/**
+	 * Add custom admin bar colors while running a localhost server.
+	 *
+	 * @return void
+	 */
+	public function custom_local_admin_bar_css() {
+
+		if ( is_admin_bar_showing() ) {
+			?>
+			<style type="text/css">
+
+			#wpadminbar #wp-admin-bar-site-name > .ab-item::after
+			{
+				content: " - localhost";
+				font-weight: 800;
+				font-family: Monospace;
+				color: #fff;
+			}
+
+			#wpadminbar #wp-admin-bar-site-name > .ab-item
+			{
+				background-color: #008000;
+				color: #ff0;
+			}
+
+			#wpadminbar #wp-admin-bar-site-name > .ab-item::before
+			{
+				background: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGFyaWEtaGlkZGVuPSJ0cnVlIiB3aWR0aD0iMTQiIGhlaWdodD0iMTYiIHN0eWxlPSItbXMtdHJhbnNmb3JtOnJvdGF0ZSgzNjBkZWcpOy13ZWJraXQtdHJhbnNmb3JtOnJvdGF0ZSgzNjBkZWcpO3RyYW5zZm9ybTpyb3RhdGUoMzYwZGVnKSI+PHBhdGggZmlsbC1ydWxlPSJldmVub2RkIiBkPSJNOS41IDNMOCA0LjUgMTEuNSA4IDggMTEuNSA5LjUgMTMgMTQgOCA5LjUgM3ptLTUgMEwwIDhsNC41IDVMNiAxMS41IDIuNSA4IDYgNC41IDQuNSAzeiIgZmlsbD0iI2ZmMCIvPjwvc3ZnPg==) center center no-repeat !important;
+				content: " " !important;
+				display: block;
+				width: 16px;
+				height: 21px;
+			}
+
+			@media screen and (max-width:782px) {
+				#wpadminbar #wp-admin-bar-site-name > .ab-item::before
+				{
+					background-size: 60% !important;
+					width: 100%;
+					height: 100%;
+					top: 0;
+				}
+			}
+
+			</style>
+			<?php
+		}
 	}
 }