+++ CHANGES.md
@@ -1,5 +1,13 @@
 #### [unreleased]
 
+#### 2.5.0 / 2020-01-23
+* update `Requires at least` in plugin file
+* add VCS checkouts automatically, thanks @Raruto
+* add Git Host icons as default setting for plugins and themes
+* add localhost admin bar coloring when on localhost, default is on
+* some now settings only display when running in localhost
+* add repositories that are added from [GitHub Updater Additions](https://github.com/afragen/github-updater-additions)
+
 #### 2.4.1 / 2019-05-09
 * a11y updates for settings tabs